Separate inputs into ApiInput component
